Spells that are actually new moved to the "new spells" section, in light of the number of new spells being added.
This is intended to evoke a more "classic look and feel" for the original spell lists - at the drawback of players now needing to go back and forth more often between the new and the old spell lists.
In the end, this was deemed to be inevitable, lest the new content eventually completely overwhelm the old.

Some "new" spells were left on the original spell lists. These were spells that would have left empty spell levels on removal and/or were deemed fundamental to the workings of the class.
"Fortune of the Gods" (the only 9th-level Luck domain spell) is an example of the former, the new "Energy Ball" is an example of the latter.